+++ SleepHunter/Models/ObservableObject.cs
@@ -1,0 +1,41 @@
+﻿using System.Collections.Generic;
+using System.ComponentModel;
+using System.Runtime.CompilerServices;
+
+namespace SleepHunter.Models
+{
+    public abstract class ObservableObject : INotifyPropertyChanged, INotifyPropertyChanging
+    {
+        public event PropertyChangedEventHandler PropertyChanged;
+        public event PropertyChangingEventHandler PropertyChanging;
+
+        protected bool SetProperty<T>(ref T backingField, T value, [CallerMemberName] string propertyName = "")
+        {
+            if (EqualityComparer<T>.Default.Equals(backingField, value))
+            {
+                return false;
+            }
+
+            OnPropertyChanging(propertyName);
+            backingField = value;
+            OnPropertyChanged(propertyName);
+            return true;
+        }
+
+        protected void OnPropertyChanged([CallerMemberName] string propertyName = "")
+        {
+            if (PropertyChanged != null)
+            {
+                PropertyChanged.Invoke(this, new PropertyChangedEventArgs(propertyName));
+            }
+        }
+
+        protected void OnPropertyChanging([CallerMemberName] string propertyName = "")
+        {
+            if (PropertyChanging != null)
+            {
+                PropertyChanging.Invoke(this, new PropertyChangingEventArgs(propertyName));
+            }
+        }
+    }
+}