+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Support for Dark Ages Client 7.41
+- Refactor status window and updating of UI
+- Refactor client memory reading for 7.41
+- Optimize some graphics allocations
 - Removed the long-defunct `Chat` window menu
 - Increased font size
 - Updated to .NET Framework 4.8 (from 3.5)